NejakaMapa13.6.17

+++ NejakaMapa/Models/Souradnice.cs
@@ -8,7 +8,7 @@ namespace NejakaMapa.Models
     public class Souradnice
     {
         public Guid Id { get; set; }
-        public double Lat { get; set; }
-        public double Lng { get; set; }
+        public decimal Lat { get; set; }
+        public decimal Lng { get; set; }
     }
 }
