+++ NejakaMapa/Models/Udalosti.cs
@@ -10,7 +10,7 @@ namespace NejakaMapa.Models
         private static List<Udalost> udalosti = new List<Udalost>();
 
 
-        public static void AddUdalost(double lat, double lng, string name, DateTime date, string text) {
+        public static void AddUdalost(decimal lat, decimal lng, string name, DateTime date, string text) {
 
             Udalost udalost = new Udalost();
             Souradnice souradnice = new Souradnice();
